Extract predictive maneuvering into new module
